+++ tests/test_diffraction_diffraction_pattern.py
@@ -237,3 +237,21 @@ class TestDiffractionPattern:
         dp = freud.diffraction.DiffractionPattern()
         with pytest.raises(ValueError):
             dp.compute((box, points))
+
+    def test_weighted_diffraction(self):
+        box = freud.box.Box(Lx=10, Ly=10, Lz=10)
+        N = 40
+        positions = np.random.uniform(-5, 5, (N, 3))
+        weights = np.random.uniform(0, 10, N)
+        dp = freud.diffraction.DiffractionPattern(grid_size=256)
+        weighted_diffraction = dp.compute((box, positions), weights=weights, reset=True)
+
+        assert np.allclose(weighted_diffraction.diffraction[128, 128], len(positions))
+
+        dp = freud.diffraction.DiffractionPattern(grid_size=256)
+
+        unweighted_diffraction = dp.compute((box, positions), reset=True)
+
+        assert (
+            unweighted_diffraction.diffraction != weighted_diffraction.diffraction
+        ).any()